Update to 20100809 release.
byuu says:

This adds some sync.sh improvements to make it handle errors more
gracefully.
It also updates asnes a good bit. All of the four base processors now
have all publicly accessible functions right at the top of the main
headers, and everything else is private. This is to allow these headers
to essentially take the place of the previous base classes in the old
bsnes-merged format. So if there's something public there, you need to
implement that exact function to make your own module.
I removed the frame counter from the PPU, as it has nothing to do with
emulation. That now resides inside the Qt -> SNES interface code. Quite
amazing, I was actually saving the frame counter into the save state
files before, yuck.
Removed some baggage in the System class: it was friending a bunch of
long-dead functions and classes.
Forgot to re-add the CHEAT_SYSTEM define to info.hpp, so that's been put
back.

+++ sync.sh
@@ -1,10 +1,13 @@
-rm -r libco
-rm -r nall
-rm -r ruby
+synchronize() {
+  if [ -d ../"$1" ]; then
+    test -d "$1" && rm -r "$1"
+    cp -r ../"$1" ./"$1"
+  fi
+}
 
-cp -r ../libco ./libco
-cp -r ../nall ./nall
-cp -r ../ruby ./ruby
+synchronize "libco"
+synchronize "nall"
+synchronize "ruby"
 
-rm -r libco/doc
-rm -r libco/test
+test -d libco/doc && rm -r libco/doc
+test -d libco/test && rm -r libco/test
